\name{DataSetsUni-package}
\alias{DataSetsUni-package}
\docType{package}
\title{
A collection of widely used univariate data sets
}

\description{
A collection of widely used univariate data sets of various applied domains on applications of distribution theory. The functions allow researchers and practitioners to quickly, easily, and efficiently access and use these data sets. The data are related to different applied domains and as follows: Bio-medical, survival analysis, medicine, reliability analysis, hydrology, actuarial science, operational research, meteorology, extreme values, quality control, engineering, finance, sports and economics. The total 100 data sets are documented along with associated references for further details and uses.
}

\details{
\tabular{ll}{
Package: \tab DataSetsUni\cr
Type: \tab Package\cr
Version: \tab 0.1 \cr
Date: \tab 2023-05-10\cr
License: \tab GPL-2\cr
}
}

\author{
Muhammad Imran <imranshakoor84@yahoo.com>, M.H Tahir <mht@iub.edu.pk> and Farrukh Jamal <farrukh.jamal@iub.edu.pk>.
}

\section{Maintainers}{
Muhammad Imran <imranshakoor84@yahoo.com>
}





\keyword{Distribution theory}
\keyword{Generalized classes}
\keyword{Compunding}
\keyword{Extended family of distributions}



