\name{Reddit Advertising}
\alias{data_reddit}
\title{The data set consists of 150 observations and is related to the Reddit advertising

}
\usage{
data_reddit
}

\description{
The function allows to provide the 150 observations and is related to the Reddit advertising.
}
\arguments{
\item{data_reddit}{
A vector of (non-negative integer) values.
}
}
\details{The data set consists of 150 observations and is related to the Reddit advertising.
Recently, it is used by Shen et al. (2022) and fitted a new generalized rayleigh distribution.
}

\value{
data_reddit gives the 150 observations and is related to the Reddit advertising.
}

\references{

Shen, Z., Alrumayh, A., Ahmad, Z., Abu-Shanab, R., Al-Mutairi, M., & Aldallal, R. (2022). A new generalized rayleigh distribution with
 analysis to big data of an online community. Alexandria Engineering Journal, 61(12), 11523-11535.



}


\author{
Muhammad Imran.

R implementation and documentation: Muhammad Imran \email{imranshakoor84@yahoo.com}.
}

%\note{
%%  ~~further notes~~
%}




\examples{
x<-data_reddit
summary(x)
}
